requirements.txt
Flask==2.3.3
textblob==0.17.1
nltk==3.8.1
openai==1.3.0
python-dotenv==1.0.0
plotly==5.15.0
pandas==2.0.3
numpy==1.24.3
httpx>=0.24.0